Move helpers to their own file
Makefile for helper

Remove extra newline
